+++ packages/gestalt-codemods/0.125.0-1.0.0/__testfixtures__/convert-box-roundedX-left.output.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Box } from 'gestalt';
+
+export default function RoundedLeft() {
+  return (
+    <Box shape="rounded">
+      roundedLeft
+    </Box>
+  );
+}