+++ packages/gestalt-codemods/0.125.0-1.0.0/__testfixtures__/convert-box-roundedX-right.input.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Box } from 'gestalt';
+
+export default function RoundedRight() {
+  return (
+    <Box shape="roundedRight">
+      roundedRight
+    </Box>
+  );
+}