+++ packages/gestalt-codemods/0.125.0-1.0.0/__testfixtures__/convert-box-shape-to-rounding-circle.output.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Box } from 'gestalt';
+
+export default function Circle() {
+  return (
+    <Box rounding="circle">
+      circle
+    </Box>
+  );
+}