+++ packages/gestalt-codemods/0.125.0-1.0.0/__testfixtures__/convert-mask-shape-to-rounding-square.output.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Mask } from 'gestalt';
+
+export default function Square() {
+  return (
+    <Mask rounding={0}>
+      square
+    </Mask>
+  );
+}