+++ packages/gestalt-codemods/0.125.0-1.0.0/__testfixtures__/convert-touchable-roundedX-right.output.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Touchable } from 'gestalt';
+
+export default function RoundedRight() {
+  return (
+    <Touchable onTouch={() => {}} shape="rounded">
+      roundedRight
+    </Touchable>
+  );
+}