+++ packages/gestalt-codemods/0.125.0-1.0.0/__testfixtures__/convert-touchable-roundedX-top.input.js
@@ -1,0 +1,11 @@
+// @flow
+import React from 'react';
+import { Touchable } from 'gestalt';
+
+export default function RoundedTop() {
+  return (
+    <Touchable onTouch={() => {}} shape="roundedTop">
+      roundedTop
+    </Touchable>
+  );
+}