+++ packages/gestalt-codemods/0.125.0-1.0.0/__tests__/convert-roundedX.test.js
@@ -1,0 +1,24 @@
+import { defineTest } from 'jscodeshift/dist/testUtils.js';
+
+jest.mock('../convert-roundedX', () => {
+  return Object.assign(require.requireActual('../convert-roundedX'), {
+    parser: 'flow',
+  });
+});
+
+describe('convert-roundedX', () => {
+  [
+    'convert-box-roundedX-bottom',
+    'convert-box-roundedX-left',
+    'convert-box-roundedX-right',
+    'convert-box-roundedX-top',
+    'convert-box-roundedX-circle',
+    'convert-touchable-roundedX-bottom',
+    'convert-touchable-roundedX-left',
+    'convert-touchable-roundedX-right',
+    'convert-touchable-roundedX-top',
+    'convert-touchable-roundedX-circle',
+  ].forEach(test => {
+    defineTest(__dirname, 'convert-roundedX', { quote: 'single' }, test);
+  });
+});