+++ packages/gestalt/src/Box.js
@@ -86,6 +86,7 @@ type Margin =
   | 12
   | 'auto';
 type Padding = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
+type Rounding = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 'circle' | 'pill';
 type PropType = {
   children?: React.Node,
   dangerouslySetInlineStyle?: {
@@ -190,6 +191,8 @@ type PropType = {
   minHeight?: number | string,
   minWidth?: number | string,
 
+  opacity?: 0 | 0.1 | 0.2 | 0.3 | 0.4 | 0.5 | 0.6 | 0.7 | 0.8 | 0.9 | 1,
+
   overflow?: 'visible' | 'hidden' | 'scroll' | 'scrollX' | 'scrollY' | 'auto',
 
   padding?: Padding,
@@ -209,15 +212,7 @@ type PropType = {
 
   position?: 'static' | 'absolute' | 'relative' | 'fixed',
   right?: boolean,
-  shape?:
-    | 'square'
-    | 'rounded'
-    | 'pill'
-    | 'circle'
-    | 'roundedTop'
-    | 'roundedBottom'
-    | 'roundedLeft'
-    | 'roundedRight',
+  rounding?: Rounding,
   shrink?: boolean,
   top?: boolean,
   width?: number | string,
@@ -245,6 +240,21 @@ const transformNumberOrPassthrough = (selector: string) => (
 
   if (m === 'auto') {
     return fromClassName(whitespace[`${selector}Auto`]);
+  }
+
+  return identity();
+};
+const rounding = (r: Rounding): Style => {
+  if (typeof r === 'number') {
+    return bind(range('rounding'), borders)(r);
+  }
+
+  if (r === 'circle') {
+    return fromClassName(borders.circle);
+  }
+
+  if (r === 'pill') {
+    return fromClassName(borders.pill);
   }
 
   return identity();
@@ -306,6 +316,14 @@ const mdPadding = union(mdPaddingX, mdPaddingY);
 const lgPaddingX = bind(rangeWithoutZero('lgPaddingX'), whitespace);
 const lgPaddingY = bind(rangeWithoutZero('lgPaddingY'), whitespace);
 const lgPadding = union(lgPaddingX, lgPaddingY);
+
+const opacityMap = mapClassName(name => styles[name]);
+const opacity = val => {
+  if (val > 0 && val < 1) {
+    return opacityMap(range('opacity0')(val * 10));
+  }
+  return opacityMap(range('opacity')(val));
+};
 
 /*
 
@@ -586,6 +604,7 @@ const propToFn = {
   maxWidth: maxWidth => fromInlineStyle({ maxWidth }),
   minHeight: minHeight => fromInlineStyle({ minHeight }),
   minWidth: minWidth => fromInlineStyle({ minWidth }),
+  opacity,
   overflow: mapping({
     hidden: layout.overflowHidden,
     scroll: layout.overflowScroll,
@@ -630,16 +649,7 @@ const propToFn = {
     // default: static
   }),
   right: toggle(layout.right0),
-  shape: mapping({
-    circle: borders.circle,
-    pill: borders.pill,
-    rounded: borders.rounded,
-    roundedBottom: borders.roundedBottom,
-    roundedLeft: borders.roundedLeft,
-    roundedRight: borders.roundedRight,
-    roundedTop: borders.roundedTop,
-    // default: square
-  }),
+  rounding,
   top: toggle(layout.top0),
   width: width => fromInlineStyle({ width }),
   wrap: toggle(layout.flexWrap),
@@ -788,6 +798,20 @@ const PaddingPropType = PropTypes.oneOf([
   10,
   11,
   12,
+]);
+
+const RoundingPropType = PropTypes.oneOf([
+  0,
+  1,
+  2,
+  3,
+  4,
+  5,
+  6,
+  7,
+  8,
+  'circle',
+  'pill',
 ]);
 
 // $FlowIssue https://github.com/facebook/flow/issues/7484
@@ -976,6 +1000,8 @@ Box.propTypes = {
   minHeight: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   minWidth: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
 
+  opacity: PropTypes.oneOf([0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1]),
+
   overflow: PropTypes.oneOf([
     'visible',
     'hidden',
@@ -1003,16 +1029,7 @@ Box.propTypes = {
 
   position: PropTypes.oneOf(['static', 'absolute', 'relative', 'fixed']),
   right: PropTypes.bool,
-  shape: PropTypes.oneOf([
-    'square',
-    'rounded',
-    'pill',
-    'circle',
-    'roundedTop',
-    'roundedBottom',
-    'roundedLeft',
-    'roundedRight',
-  ]),
+  rounding: RoundingPropType,
   top: PropTypes.bool,
   width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   wrap: PropTypes.bool,