+++ packages/gestalt/src/Tooltip.js
@@ -46,7 +46,6 @@ export default class Tooltip extends React.Component<Props, State> {
           <Controller
             anchor={anchor}
             bgColor="darkGray"
-            caret={false}
             idealDirection="down"
             onDismiss={noop}
             positionRelativeToAnchor